requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-htmx>=1.17
python-dotenv>=1.0
gunicorn>=21.0
mysqlclient>=2.2
django-markdownx>=4.0
markdown>=3.5
